backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
PyMySQL==1.1.0
python-dotenv==1.0.0
PyPDF2==3.0.1
pdfplumber==0.10.3
Pillow>=10.0.0
sentence-transformers>=2.2.0
numpy>=1.24.0
scikit-learn>=1.3.0
torch>=2.0.0
transformers>=4.30.0
python-multipart==0.0.6
google-generativeai>=0.3.0
bcrypt>=4.0.0
Flask-Mail>=0.9.1
